remove unneeded step, component
