+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/dataset/DatasetCloneInputs.java
@@ -1,20 +1,16 @@
 package bio.terra.workspace.service.resource.controlled.flight.clone.dataset;
 
-import java.util.List;
 import java.util.UUID;
 
 public class DatasetCloneInputs {
   private final UUID workspaceId;
   private final String projectId;
   private final String datasetName;
-  private final List<String> roleNames;
 
-  public DatasetCloneInputs(
-      UUID workspaceId, String projectId, String datasetName, List<String> roleNames) {
+  public DatasetCloneInputs(UUID workspaceId, String projectId, String datasetName) {
     this.workspaceId = workspaceId;
     this.projectId = projectId;
     this.datasetName = datasetName;
-    this.roleNames = roleNames;
   }
 
   public UUID getWorkspaceId() {
@@ -29,10 +25,6 @@ public class DatasetCloneInputs {
     return datasetName;
   }
 
-  public List<String> getRoleNames() {
-    return roleNames;
-  }
-
   @Override
   public String toString() {
     return "DatasetCloneInputs{"
@@ -44,8 +36,6 @@ public class DatasetCloneInputs {
         + ", datasetName='"
         + datasetName
         + '\''
-        + ", roleNames="
-        + roleNames
         + '}';
   }
